allow alternatives to w3m to browse docs, add backup url
-b,--backup causes requests to use mercury-in.space instead of mercurylang.org
(needed it myself as wget and w3m were hanging on trying the IPV6 connection).

$WWWPAGER / w3m / xdg-open tried in turn for a browser.
not using xdg-open first as that'll usually be a graphical browser.

minor refactoring with browse_to/3. we're still using a shell to start a
webbrowser, but it'll be very easy now to skip the shell.