Added support for --rotd option, to source ROTD remote docs.
Added a new command line option, "-r, --rotd" and the supporting
code to make it work. This involved changes to ensure_cache and
refresh_cache to not be fixed on the one source, formerly just
the Backup option, a new type was added to allow this.

Tested, all existing options work fine, as does --rotd / -r.